+++ packages/core/src/components/viewer-grid/viewer-grid.stories.ts
@@ -1,6 +1,9 @@
-import { storiesOfViewerGrid } from "@dso-toolkit/sources";
+import { storiesOfViewerGrid, Tile } from "@dso-toolkit/sources";
 import { storiesOf } from "@storybook/web-components";
 import { html } from "lit-html";
+
+import { tileGridTemplate } from '../../../../../packages/css/src/components/tile-grid/tile-grid.template';
+import { tileTemplate } from '../../../../../packages/css/src/components/tile/tile.template';
 
 import { anchorTemplate } from "../anchor/anchor.template";
 import { badgeTemplate } from "../badge/badge.template";
@@ -11,7 +14,7 @@ import { viewerGridDocumentListItem } from "./viewer-grid-document-list-item.tem
 import { viewerGridFilterblok } from "./viewer-grid-filterblok.template";
 import { viewerGridTemplate } from "./viewer-grid.template";
 
-function viewerGridMainDemoTemplate(
+function viewerGridWithSearchResultsDemoTemplate(
   filterblokAllOptions: (e: MouseEvent) => void,
   filterblokDeleteActiveFilter: (e: MouseEvent) => void,
   documentHeaderFeatureAction: (e: MouseEvent) => void,
@@ -151,6 +154,14 @@ function viewerGridMainDemoTemplate(
   `;
 }
 
+function viewerGridWithTilesDemoTemplate(tiles: Tile[]) {
+  return html`
+    <h2>Thema's</h2>
+    <p>Bekijk regels en beleid rond een bepaald thema.</p>
+    ${tileGridTemplate({ children: html`${tiles.map(tile => tileTemplate(tile))}` })}
+  `;
+}
+
 const filterpanel = html`
   <hr>
   <dso-alert status="info">
@@ -266,7 +277,7 @@ storiesOfViewerGrid(
     readme,
   },
   {
-    viewerGridDemoTemplate: ({
+    viewerGridWithSearchResultsDemoTemplate: ({
       closeOverlay,
       filterpanelOpen,
       overlayOpen,
@@ -287,7 +298,7 @@ storiesOfViewerGrid(
         filterpanelApply,
         filterpanelCancel,
         noOverlay,
-        main: viewerGridMainDemoTemplate(
+        main: viewerGridWithSearchResultsDemoTemplate(
           allOptions,
           filterblokDeleteActiveFilter,
           documentHeaderFeatureAction,
@@ -295,7 +306,28 @@ storiesOfViewerGrid(
           documentHeaderFeaturesOpen
         ),
         map,
-        overlay,
+        overlay
+      }),
+    viewerGridWithTilesDemoTemplate: ({
+      closeOverlay,
+      tiles,
+      filterpanelOpen,
+      overlayOpen,
+      noOverlay,
+      filterpanelApply,
+      filterpanelCancel,
+    }) =>
+      viewerGridTemplate({
+        closeOverlay,
+        filterpanelOpen,
+        filterpanel,
+        filterpanelApply,
+        filterpanelCancel,
+        overlayOpen,
+        noOverlay,
+        main: viewerGridWithTilesDemoTemplate(tiles),
+        map,
+        overlay
       }),
   }
 );