+++ packages/css/src/components/tile/tile.template.ts
@@ -1,11 +1,12 @@
 import { Tile } from '@dso-toolkit/sources';
 import { html } from 'lit-html';
+import { classMap } from 'lit-html/directives/class-map';
 
 import { imageTemplate } from '../image/image.template';
 
-export function tileTemplate({ anchor, image, label }: Tile) {
+export function tileTemplate({ anchor, image, label, variant }: Tile) {
   return html`
-    <div class="dso-tile">
+    <div class="dso-tile ${classMap({ 'dso-theme': variant === 'theme' })}">
       <a href=${anchor}>
         <span class="dso-tile-icon">
           ${imageTemplate(image)}