+++ packages/sources/src/components/tile-grid/tile-grid.args.ts
@@ -1,0 +1,7 @@
+import { ArgTypes } from '../../stories-helpers';
+
+export interface TileGridArgs {
+}
+
+export const tileGridArgTypes: ArgTypes<TileGridArgs> = {
+};