+++ packages/sources/src/components/tile-grid/tile-grid.models.ts
@@ -1,0 +1,3 @@
+export interface TileGrid<T> {
+  children?: T;
+}