+++ packages/sources/src/components/tile/tile.args.ts
@@ -6,6 +6,7 @@ export interface TileArgs {
   label: string;
   imageSource: string;
   imageAlt: string;
+  variant?: 'theme';
 }
 
 export const tileArgTypes: ArgTypes<TileArgs> = {
@@ -23,6 +24,16 @@ export const tileArgTypes: ArgTypes<TileArgs> = {
     control: {
       type: 'text'
     }
+  },
+  variant: {
+    options: [undefined, 'theme'],
+    control: {
+      type: 'select',
+      labels: {
+        undefined: 'default',
+        'theme': 'theme'
+      }
+    }
   }
 };
 
@@ -33,6 +44,7 @@ export function tileArgsMapper(a: TileArgs): Tile {
     image: {
       source: a.imageSource,
       alt: a.imageAlt
-    }
+    },
+    variant: a.variant
   };
 }