+++ packages/sources/src/components/viewer-grid/viewer-grid.models.ts
@@ -1,17 +1,24 @@
-import { DefinitionList, Label } from "../..";
+import { DefinitionList, Label, Tile } from "../..";
 
-export interface ViewerGridDemoProperties {
+interface ViewerGridDemoProperties {
   filterpanelOpen: boolean;
   overlayOpen: boolean;
   noOverlay: boolean;
   closeOverlay: (e: MouseEvent) => void;
   filterpanelCancel: (e: MouseEvent) => void;
   filterpanelApply: (e: MouseEvent) => void;
+}
+
+export interface ViewerGridWithSearchResultsProperties extends ViewerGridDemoProperties {
   allOptions: (e: MouseEvent) => void;
   filterblokDeleteActiveFilter: (e: MouseEvent) => void;
   documentHeaderFeaturesOpen: boolean;
   documentHeaderMapAction: (e: MouseEvent) => void;
   documentHeaderFeatureAction: (e: MouseEvent) => void;
+}
+
+export interface ViewerGridWithTilesProperties extends ViewerGridDemoProperties {
+  tiles: Tile[];
 }
 
 export interface ViewerGrid<TemplateFnReturnType> {